Refactor logic for uploading a single file into function.
Do not use files_to_upload, instead process each file as it is visited during the directory traversal.
Touch an empty *.uploaded file to mark files that have been uploaded (as a fallback in case the script exits without upgrading the upload record)
Add Mac notifications (terminal-notifier) for when a folder has been uploaded, when a file fails to upload, or when the script exits due to an exception.